refactor(chain_manager): extract high level transaction factory
Into its own file.
Create some helper methods: get_pkh and insert_change_output
