+++ node/src/actors/chain_manager/mod.rs
@@ -43,11 +43,7 @@ use crate::actors::{
     sessions_manager::SessionsManager,
     storage_keys::CHAIN_STATE_KEY,
 };
-use crate::{signature_mngr, storage_mngr};
-use futures::Future;
-use witnet_data_structures::chain::{
-    DataRequestOutput, Input, TransactionBody, ValueTransferInput, ValueTransferOutput,
-};
+use crate::storage_mngr;
 use witnet_data_structures::{
     chain::{
         Block, ChainState, CheckpointBeacon, DataRequestReport, Epoch, Hash, Hashable,
@@ -63,6 +59,8 @@ use witnet_validations::validations::{validate_block, validate_candidate, Diff};
 mod actor;
 mod handlers;
 mod mining;
+/// High level transaction factory
+pub mod transaction_factory;
 
 /// Maximum blocks number to be sent during synchronization process
 pub const MAX_BLOCKS_SYNC: usize = 500;
@@ -393,137 +391,6 @@ fn update_pools(
 
     utxo_diff.apply(unspent_outputs_pool);
 }
-
-// TODO: move to transaction_factory.rs
-/// Select enough UTXOs to sum up to `amount`.
-///
-/// On success, return a list of output pointers and their sum.
-/// On error, return the total sum of the output pointers in `own_utxos`.
-pub fn take_enough_utxos<S: std::hash::BuildHasher>(
-    own_utxos: &HashSet<OutputPointer, S>,
-    all_utxos: &UnspentOutputsPool,
-    amount: u64,
-) -> Result<(Vec<OutputPointer>, u64), u64> {
-    let mut acc = 0;
-    let mut list = vec![];
-
-    for op in own_utxos {
-        acc += all_utxos[op].value();
-        list.push(op.clone());
-        if acc >= amount {
-            break;
-        }
-    }
-
-    if acc >= amount {
-        Ok((list, acc))
-    } else {
-        Err(acc)
-    }
-}
-
-/// Build value transfer transaction with the given outputs and fee.
-pub fn build_vtt<S: std::hash::BuildHasher>(
-    mut outputs: Vec<ValueTransferOutput>,
-    fee: u64,
-    own_utxos: &HashSet<OutputPointer, S>,
-    own_pkh: PublicKeyHash,
-    all_utxos: &UnspentOutputsPool,
-) -> Result<TransactionBody, u64> {
-    let output_value: u64 = outputs.iter().map(|x| x.value).sum();
-    match take_enough_utxos(own_utxos, all_utxos, output_value + fee) {
-        Err(sum_of_own_utxos) => Err(sum_of_own_utxos),
-        Ok((output_pointers, input_value)) => {
-            let change_amount = input_value - output_value - fee;
-            if change_amount > 0 {
-                // Create change output
-                outputs.push(ValueTransferOutput {
-                    pkh: own_pkh,
-                    value: change_amount,
-                });
-            }
-
-            let inputs = output_pointers
-                .into_iter()
-                .map(|x| {
-                    Input::ValueTransfer(ValueTransferInput {
-                        transaction_id: x.transaction_id,
-                        output_index: x.output_index,
-                    })
-                })
-                .collect();
-            let outputs = outputs.into_iter().map(Output::ValueTransfer).collect();
-            let tx_body = TransactionBody::new(0, inputs, outputs);
-            Ok(tx_body)
-        }
-    }
-}
-
-/// Build data request transaction with the given outputs and fee.
-pub fn build_drt<S: std::hash::BuildHasher>(
-    dr: DataRequestOutput,
-    fee: u64,
-    own_utxos: &HashSet<OutputPointer, S>,
-    own_pkh: PublicKeyHash,
-    all_utxos: &UnspentOutputsPool,
-) -> Result<TransactionBody, u64> {
-    let dro = Output::DataRequest(dr);
-    let output_value: u64 = dro.value();
-    match take_enough_utxos(own_utxos, all_utxos, output_value + fee) {
-        Err(sum_of_own_utxos) => Err(sum_of_own_utxos),
-        Ok((output_pointers, input_value)) => {
-            let mut outputs = vec![];
-            let change_amount = input_value - output_value - fee;
-            if change_amount > 0 {
-                // Create change output
-                outputs.push(Output::ValueTransfer(ValueTransferOutput {
-                    pkh: own_pkh,
-                    value: change_amount,
-                }));
-            }
-            outputs.push(dro);
-
-            let inputs = output_pointers
-                .into_iter()
-                .map(|x| {
-                    Input::ValueTransfer(ValueTransferInput {
-                        transaction_id: x.transaction_id,
-                        output_index: x.output_index,
-                    })
-                })
-                .collect();
-            let tx_body = TransactionBody::new(0, inputs, outputs);
-            Ok(tx_body)
-        }
-    }
-}
-
-/// Sign a transaction using this node's private key.
-/// This function assumes that all the inputs have the same public key hash:
-/// the hash of the public key of the node.
-pub fn sign_transaction(
-    tx_body: TransactionBody,
-) -> impl Future<Item = Transaction, Error = failure::Error> {
-    // Assuming that all the inputs have the same pkh
-    signature_mngr::sign(&tx_body).map(move |signature| {
-        // TODO: do we need to sign:
-        // value transfer inputs,
-        // data request inputs (for commits),
-        // commit inputs (for reveals),
-        //
-        // We do not need to sign:
-        // reveal inputs (for tallies)
-        //
-        // But currently we just sign everything, hoping that the validations
-        // work
-        let num_inputs = tx_body.inputs.len();
-        let signatures = vec![signature; num_inputs];
-
-        Transaction::new(tx_body, signatures)
-    })
-}
-
-// TODO: END
 
 fn show_info_tally(
     unspent_outputs_pool: &UnspentOutputsPool,
